Make the compact matrix more compact

+++ include/common/compact_matrix.h
@@ -3,11 +3,12 @@
 
 #include <assert.h>
 #include <stddef.h>
+#include <stdint.h>
 
 typedef struct {
     double value;
-    size_t row;
-    size_t column;
+    uint32_t row;
+    uint32_t column;
 } Item;
 
 typedef struct {
@@ -20,12 +21,12 @@ typedef struct {
     size_t n_cols;
 } CompactMatrix;
 
-CompactMatrix cmatrix_make(size_t rows, size_t columns, size_t num_elems);
+CompactMatrix cmatrix_make(uint32_t rows, uint32_t columns, size_t num_elems);
 
 CompactMatrix
-cmatrix_make_without_lengths(size_t rows, size_t columns, size_t num_elems);
+cmatrix_make_without_lengths(uint32_t rows, uint32_t columns, size_t num_elems);
 
-void cmatrix_add(CompactMatrix* m, size_t row, size_t column, double value);
+void cmatrix_add(CompactMatrix* m, uint32_t row, uint32_t column, double value);
 
 void cmatrix_sort(CompactMatrix* m);
